Refactoring - eliminating useless AbstractServiceBuilder class
